@@ -1,7 +1,44 @@
-# This file should contain all the record creation needed to seed the database with its default values.
-# The data can then be loaded with the bin/rails db:seed command (or created alongside the database with db:setup).
-#
-# Examples:
-#
-#   movies = Movie.create([{ name: 'Star Wars' }, { name: 'Lord of the Rings' }])
-#   Character.create(name: 'Luke', movie: movies.first)
+Category.create(
+  category_name: "Handmade items"
+)
+
+Category.create(
+  category_name: "Destash Yarn"
+)
+
+
+Item.create(
+  item_name: "Large purple triangular shawl",
+  price: 180.00,
+  description: "Large hand knit shawl made using hand dyed yarns and brioche knitting techniques",
+  quantity_available: 1,
+  photo_url: "https://res.cloudinary.com/dpqgbq6bo/image/upload/v1686169873/IMG_4956_rbzkyh.jpg",
+  category_id: 1
+)
+
+Item.create(
+  item_name: "Multi-colored scarf",
+  price: 200.00,
+  description: "Beautiful hand knit scarf made using plethora of hand dyed mini skeins",
+  quantity_available: 1,
+  photo_url:"https://res.cloudinary.com/dpqgbq6bo/image/upload/v1686169864/IMG_4949_pns5pg.jpg",
+  category_id: 1
+)
+
+Item.create(
+  item_name: "Green Crocheted Llama",
+  price: 30.00,
+  description: "Hand crocheted amigarumi llama",
+  quantity_available: 1,
+  photo_url: "https://res.cloudinary.com/dpqgbq6bo/image/upload/v1686169864/IMG_4843_j1zck8.jpg",
+  category_id: 1
+)
+
+Item.create(
+  item_name: "Large blue triangular shawl",
+  price: 180.00,
+  description: "Large hand knit shawl made using hand dyed yarns and brioche knitting techniques",
+  quantity_available: 1,
+  photo_url: "https://res.cloudinary.com/dpqgbq6bo/image/upload/v1686169869/IMG_4935_uhsrnd.jpg",
+  category_id: 1
+)
